Baggage Search optimization, adiing of toast
