+++ mobile/app/_layout.tsx
@@ -1,6 +1,8 @@
 import { Stack } from "expo-router";
 import "../global.css";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
+import Toast from "react-native-toast-message";
+import { toastConfig } from "../styles/toastStyles";
 // import { StatusBar } from "expo-status-bar";
 
 const queryClient = new QueryClient();
@@ -11,6 +13,7 @@ export default function RootLayout() {
         <Stack screenOptions={{ headerShown: false }}>
           <Stack.Screen name="(tabs)" />
         </Stack>
+        <Toast config={toastConfig} />
       </QueryClientProvider>
   );
 }
