+++ mobile/components/baggage/ReportBaggageModal.tsx
@@ -1,13 +1,13 @@
-import { Text, View, ScrollView, Modal, Alert, TouchableOpacity, ActivityIndicator, Dimensions, Platform } from "react-native";
+import { Text, View, ScrollView, Modal, Alert, TouchableOpacity, TextInput, ActivityIndicator, Dimensions, Platform } from "react-native";
 import React, { useMemo, useState } from "react";
 import { useSafeAreaInsets } from "react-native-safe-area-context";
 
 
-import { BAGGAGE_TYPES, TRANSPORT_TYPES, ROUTE_TYPES, PROVINCES, PROVINCE_DISTRICT_MAP } from "@/constants/allowedValues";
-import { OptionPicker, SelectField, toTitleCase } from "./FormsHelper";
+import { BAGGAGE_TYPES, TRANSPORT_TYPES, ROUTE_TYPES, PROVINCES, PROVINCE_DISTRICT_MAP, PHONE_NUMBER_REGEX, PHONE_EXAMPLE } from "@/constants/allowedValues";
+import { OptionPicker, SelectField, toTitleCase } from "../FormsHelper";
 import { showAlerts } from "@/utils/alerts";
 
-interface SearchBaggageModalProps {
+interface ReportBaggageModalProps {
     isVisible: boolean;
     onClose: () => void;
     formData: {
@@ -16,15 +16,17 @@ interface SearchBaggageModalProps {
         routeType: string;
         destinationProvince: string;
         destinationDistrict: string;
+        destination: string;
+        docLocation: string;
+        finderContact: string;
     };
-    searchBaggage: (params: any) => void;
+    reportBaggage: () => void;
     updateFormData: (field: string, value: string) => void;
-    isSearching: boolean;
-    resetSearch: () => void;
+    isReporting: boolean;
 }
 
 
-const SearchBaggageModal = ({ isVisible, onClose, formData, searchBaggage, updateFormData, isSearching, resetSearch }: SearchBaggageModalProps) => {
+const ReportBaggageModal = ({ isVisible, onClose, formData, reportBaggage, updateFormData, isReporting }: ReportBaggageModalProps) => {
     const [openPicker, setOpenPicker] = useState<null | 'baggageType' | 'transportType' | 'routeType' | 'destinationProvince' | 'destinationDistrict'>(null);
 
     const insets = useSafeAreaInsets();
@@ -34,22 +36,28 @@ const SearchBaggageModal = ({ isVisible, onClose, formData, searchBaggage, updat
         return prov ? (PROVINCE_DISTRICT_MAP[prov] ?? []) : [];
     }, [formData.destinationProvince]);
 
+    const isPhoneValid = PHONE_NUMBER_REGEX.test(formData.finderContact);
     const isFormComplete = Boolean(
         formData.baggageType &&
         formData.transportType &&
         formData.routeType &&
         formData.destinationProvince &&
-        formData.destinationDistrict
+        formData.destinationDistrict &&
+        formData.destination &&
+        formData.docLocation &&
+        formData.finderContact
     );
 
-    const handleSearch = () => {
+    const handleSave = () => {
         if (!isFormComplete) {
-            showAlerts("Error", "Please fill in all required fields.");
+            Alert.alert("Error", "Please fill in all required fields.");
             return;
         }
-        // Reset previous search results before searching
-        resetSearch();
-        searchBaggage(formData);
+        if (!isPhoneValid) {
+            Alert.alert("Error", `Invalid phone number format. Example: ${PHONE_EXAMPLE}`);
+            return;
+        }
+        reportBaggage();
     };
 
     const renderSelect = (label: string, value: string, onPress: () => void, placeholder: string, displayValue?: string) => (
@@ -72,19 +80,23 @@ const SearchBaggageModal = ({ isVisible, onClose, formData, searchBaggage, updat
                     <TouchableOpacity onPress={() => {
                         showAlerts("Cancel", "Are you sure you want to cancel?", [
                             { text: "No", style: "cancel" },
-                            { text: "Yes", style: "destructive", onPress: onClose },
+                            {
+                                text: "Yes",
+                                style: "destructive",
+                                onPress: onClose,
+                            },
                         ]);
                     }}>
                         <Text className="text-red-500 font-semibold">Close</Text>
                     </TouchableOpacity>
-                    <TouchableOpacity onPress={handleSearch} disabled={isSearching || !isFormComplete }>
-                        {(!isFormComplete) ? (
-                            <Text className="text-gray-500 font-semibold">Search</Text>
+                    <TouchableOpacity onPress={handleSave} disabled={isReporting || !isFormComplete || !isPhoneValid}>
+                        {(!isFormComplete || !isPhoneValid) ? (
+                            <Text className="text-gray-500 font-semibold">Upload</Text>
                         ) :
-                            isSearching ? (
+                            isReporting ? (
                                 <ActivityIndicator size="small" color="blue" />
                             ) : (
-                                <Text className="text-blue-500 font-bold">Search</Text>
+                                <Text className="text-blue-500 font-bold">Upload</Text>
                             )}
                     </TouchableOpacity>
                 </View>
@@ -103,7 +115,7 @@ const SearchBaggageModal = ({ isVisible, onClose, formData, searchBaggage, updat
                         'Baggage Type', // Title of the select field, as per the form requirements in Baggage.Helpers.tsx
                         formData.baggageType, // Current selected value for baggage type
                         () => setOpenPicker('baggageType'), // Function to open the baggage type picker
-                        'How would you classify the baggage you lost?' // Placeholder text when no value is selected
+                        'How would you classify the baggage you found?' // Placeholder text when no value is selected
                     )}
 
                     {renderSelect(
@@ -117,7 +129,7 @@ const SearchBaggageModal = ({ isVisible, onClose, formData, searchBaggage, updat
                         'Route Type',
                         formData.routeType,
                         () => setOpenPicker('routeType'),
-                        'Were you traveling within the same area / town (local) or from one town to another (intercity)?'
+                        'Where you traveling within the same area / town (local) or from one town to another (intercity)?'
                     )}
 
                     {renderSelect(
@@ -136,13 +148,51 @@ const SearchBaggageModal = ({ isVisible, onClose, formData, searchBaggage, updat
                         disabled={!formData.destinationProvince}
                     />
 
+                    <Text className="text-lg font-semibold text-gray-600 mb-2">Destination</Text>
+                    <TextInput
+                        className="border border-gray-300 rounded p-2 mb-4"
+                        placeholder='Where was the vehicle "ending its trip"?'
+                        value={formData.destination}
+                        onChangeText={(value) => updateFormData('destination', value)}
+                        multiline
+                        maxLength={100}
+                    />
+
+                    <Text className="text-lg font-semibold text-gray-600 mb-2">Baggage Location</Text>
+                    <TextInput
+                        className="border border-gray-300 rounded p-2 mb-4"
+                        placeholder="Where can the owner come to collect the baggage?"
+                        value={formData.docLocation}
+                        onChangeText={(value) => updateFormData('docLocation', value)}
+                        multiline
+                        maxLength={200}
+                    />
+
+                    <Text className="text-lg font-semibold text-gray-600 mb-2">Finder Contact</Text>
+                    <TextInput
+                        className="border border-gray-300 rounded p-2 mb-4"
+                        placeholder="On what number can the owner reach you?"
+                        value={formData.finderContact}
+                        onChangeText={(value) => {
+                            const digits = value.replace(/\D/g, '');
+                            updateFormData('finderContact', digits);
+                        }}
+                        keyboardType="number-pad"
+                        maxLength={10}
+                    />
+                    {formData.finderContact.length > 0 && !isPhoneValid && (
+                        <Text className="text-red-600 text-xs mb-7">Invalid phone number format. Example: {PHONE_EXAMPLE}</Text>
+                    )}
+                    {formData.finderContact.length === 0 && (
+                        <View className="mb-7" />
+                    )}
 
                     {/* Pickers */}
                     <OptionPicker
                         visible={openPicker === 'baggageType'}
                         title="Select Baggage Type"
                         options={BAGGAGE_TYPES as unknown as string[]}
-                        onSelect={(val) => updateFormData('baggageType', String(val).toLowerCase())}
+                        onSelect={(val: string) => updateFormData('baggageType', String(val).toLowerCase())}
                         onClose={() => setOpenPicker(null)}
                     />
 
@@ -150,7 +200,7 @@ const SearchBaggageModal = ({ isVisible, onClose, formData, searchBaggage, updat
                         visible={openPicker === 'transportType'}
                         title="Select Transport Type"
                         options={TRANSPORT_TYPES as unknown as string[]}
-                        onSelect={(val) => updateFormData('transportType', String(val).toLowerCase())}
+                        onSelect={(val: string) => updateFormData('transportType', String(val).toLowerCase())}
                         onClose={() => setOpenPicker(null)}
                     />
 
@@ -158,7 +208,7 @@ const SearchBaggageModal = ({ isVisible, onClose, formData, searchBaggage, updat
                         visible={openPicker === 'routeType'}
                         title="Select Route Type"
                         options={ROUTE_TYPES as unknown as string[]}
-                        onSelect={(val) => updateFormData('routeType', String(val).toLowerCase())}
+                        onSelect={(val: string) => updateFormData('routeType', String(val).toLowerCase())}
                         onClose={() => setOpenPicker(null)}
                     />
 
@@ -166,22 +216,22 @@ const SearchBaggageModal = ({ isVisible, onClose, formData, searchBaggage, updat
                         visible={openPicker === 'destinationProvince'}
                         title="Select Destination Province"
                         options={PROVINCES as unknown as string[]}
-                        onSelect={(val) => {
+                        onSelect={(val: string) => {
                             updateFormData('destinationProvince', String(val).toLowerCase());
                             // reset district when province changes
                             updateFormData('destinationDistrict', '');
                         }}
                         onClose={() => setOpenPicker(null)}
-                        getLabel={(v) => toTitleCase(v)}
+                        getLabel={(v: string) => toTitleCase(v)}
                     />
 
                     <OptionPicker
                         visible={openPicker === 'destinationDistrict'}
                         title="Select Destination District"
                         options={districts as unknown as string[]}
-                        onSelect={(val) => updateFormData('destinationDistrict', String(val).toLowerCase())}
+                        onSelect={(val: string) => updateFormData('destinationDistrict', String(val).toLowerCase())}
                         onClose={() => setOpenPicker(null)}
-                        getLabel={(v) => toTitleCase(v)}
+                        getLabel={(v: string) => toTitleCase(v)}
                     />
                 </ScrollView>
             </View>
@@ -189,4 +239,4 @@ const SearchBaggageModal = ({ isVisible, onClose, formData, searchBaggage, updat
     );
 };
 
-export default SearchBaggageModal;
+export default ReportBaggageModal;
