+++ mobile/components/baggage/SearchBaggageModal.tsx
@@ -1,13 +1,13 @@
-import { Text, View, ScrollView, Modal, Alert, TouchableOpacity, TextInput, ActivityIndicator, Dimensions, Platform } from "react-native";
+import { Text, View, ScrollView, Modal, Alert, TouchableOpacity, ActivityIndicator, Dimensions, Platform } from "react-native";
 import React, { useMemo, useState } from "react";
 import { useSafeAreaInsets } from "react-native-safe-area-context";
 
 
-import { BAGGAGE_TYPES, TRANSPORT_TYPES, ROUTE_TYPES, PROVINCES, PROVINCE_DISTRICT_MAP, PHONE_NUMBER_REGEX, PHONE_EXAMPLE } from "@/constants/allowedValues";
-import { OptionPicker, SelectField, toTitleCase } from "./FormsHelper";
+import { BAGGAGE_TYPES, TRANSPORT_TYPES, ROUTE_TYPES, PROVINCES, PROVINCE_DISTRICT_MAP } from "@/constants/allowedValues";
+import { OptionPicker, SelectField, toTitleCase } from "../FormsHelper";
 import { showAlerts } from "@/utils/alerts";
 
-interface ReportBaggageModalProps {
+interface SearchBaggageModalProps {
     isVisible: boolean;
     onClose: () => void;
     formData: {
@@ -16,17 +16,15 @@ interface ReportBaggageModalProps {
         routeType: string;
         destinationProvince: string;
         destinationDistrict: string;
-        destination: string;
-        docLocation: string;
-        finderContact: string;
     };
-    reportBaggage: () => void;
+    searchBaggage: (params: any) => void;
     updateFormData: (field: string, value: string) => void;
-    isReporting: boolean;
+    isSearching: boolean;
+    resetSearch: () => void;
 }
 
 
-const ReportBaggageModal = ({ isVisible, onClose, formData, reportBaggage, updateFormData, isReporting }: ReportBaggageModalProps) => {
+const SearchBaggageModal = ({ isVisible, onClose, formData, searchBaggage, updateFormData, isSearching, resetSearch }: SearchBaggageModalProps) => {
     const [openPicker, setOpenPicker] = useState<null | 'baggageType' | 'transportType' | 'routeType' | 'destinationProvince' | 'destinationDistrict'>(null);
 
     const insets = useSafeAreaInsets();
@@ -36,28 +34,22 @@ const ReportBaggageModal = ({ isVisible, onClose, formData, reportBaggage, updat
         return prov ? (PROVINCE_DISTRICT_MAP[prov] ?? []) : [];
     }, [formData.destinationProvince]);
 
-    const isPhoneValid = PHONE_NUMBER_REGEX.test(formData.finderContact);
     const isFormComplete = Boolean(
         formData.baggageType &&
         formData.transportType &&
         formData.routeType &&
         formData.destinationProvince &&
-        formData.destinationDistrict &&
-        formData.destination &&
-        formData.docLocation &&
-        formData.finderContact
+        formData.destinationDistrict
     );
 
-    const handleSave = () => {
+    const handleSearch = () => {
         if (!isFormComplete) {
-            Alert.alert("Error", "Please fill in all required fields.");
+            showAlerts("Error", "Please fill in all required fields.");
             return;
         }
-        if (!isPhoneValid) {
-            Alert.alert("Error", `Invalid phone number format. Example: ${PHONE_EXAMPLE}`);
-            return;
-        }
-        reportBaggage();
+        // Reset previous search results before searching
+        resetSearch();
+        searchBaggage(formData);
     };
 
     const renderSelect = (label: string, value: string, onPress: () => void, placeholder: string, displayValue?: string) => (
@@ -80,23 +72,19 @@ const ReportBaggageModal = ({ isVisible, onClose, formData, reportBaggage, updat
                     <TouchableOpacity onPress={() => {
                         showAlerts("Cancel", "Are you sure you want to cancel?", [
                             { text: "No", style: "cancel" },
-                            {
-                                text: "Yes",
-                                style: "destructive",
-                                onPress: onClose,
-                            },
+                            { text: "Yes", style: "destructive", onPress: onClose },
                         ]);
                     }}>
                         <Text className="text-red-500 font-semibold">Close</Text>
                     </TouchableOpacity>
-                    <TouchableOpacity onPress={handleSave} disabled={isReporting || !isFormComplete || !isPhoneValid}>
-                        {(!isFormComplete || !isPhoneValid) ? (
-                            <Text className="text-gray-500 font-semibold">Upload</Text>
+                    <TouchableOpacity onPress={handleSearch} disabled={isSearching || !isFormComplete }>
+                        {(!isFormComplete) ? (
+                            <Text className="text-gray-500 font-semibold">Search</Text>
                         ) :
-                            isReporting ? (
+                            isSearching ? (
                                 <ActivityIndicator size="small" color="blue" />
                             ) : (
-                                <Text className="text-blue-500 font-bold">Upload</Text>
+                                <Text className="text-blue-500 font-bold">Search</Text>
                             )}
                     </TouchableOpacity>
                 </View>
@@ -115,7 +103,7 @@ const ReportBaggageModal = ({ isVisible, onClose, formData, reportBaggage, updat
                         'Baggage Type', // Title of the select field, as per the form requirements in Baggage.Helpers.tsx
                         formData.baggageType, // Current selected value for baggage type
                         () => setOpenPicker('baggageType'), // Function to open the baggage type picker
-                        'How would you classify the baggage you found?' // Placeholder text when no value is selected
+                        'How would you classify the baggage you lost?' // Placeholder text when no value is selected
                     )}
 
                     {renderSelect(
@@ -129,7 +117,7 @@ const ReportBaggageModal = ({ isVisible, onClose, formData, reportBaggage, updat
                         'Route Type',
                         formData.routeType,
                         () => setOpenPicker('routeType'),
-                        'Where you traveling within the same area / town (local) or from one town to another (intercity)?'
+                        'Were you traveling within the same area / town (local) or from one town to another (intercity)?'
                     )}
 
                     {renderSelect(
@@ -148,44 +136,6 @@ const ReportBaggageModal = ({ isVisible, onClose, formData, reportBaggage, updat
                         disabled={!formData.destinationProvince}
                     />
 
-                    <Text className="text-lg font-semibold text-gray-600 mb-2">Destination</Text>
-                    <TextInput
-                        className="border border-gray-300 rounded p-2 mb-4"
-                        placeholder='Where was the vehicle "ending its trip"?'
-                        value={formData.destination}
-                        onChangeText={(value) => updateFormData('destination', value)}
-                        multiline
-                        maxLength={100}
-                    />
-
-                    <Text className="text-lg font-semibold text-gray-600 mb-2">Baggage Location</Text>
-                    <TextInput
-                        className="border border-gray-300 rounded p-2 mb-4"
-                        placeholder="Where can the owner come to collect the baggage?"
-                        value={formData.docLocation}
-                        onChangeText={(value) => updateFormData('docLocation', value)}
-                        multiline
-                        maxLength={200}
-                    />
-
-                    <Text className="text-lg font-semibold text-gray-600 mb-2">Finder Contact</Text>
-                    <TextInput
-                        className="border border-gray-300 rounded p-2 mb-4"
-                        placeholder="On what number can the owner reach you?"
-                        value={formData.finderContact}
-                        onChangeText={(value) => {
-                            const digits = value.replace(/\D/g, '');
-                            updateFormData('finderContact', digits);
-                        }}
-                        keyboardType="number-pad"
-                        maxLength={10}
-                    />
-                    {formData.finderContact.length > 0 && !isPhoneValid && (
-                        <Text className="text-red-600 text-xs mb-7">Invalid phone number format. Example: {PHONE_EXAMPLE}</Text>
-                    )}
-                    {formData.finderContact.length === 0 && (
-                        <View className="mb-7" />
-                    )}
 
                     {/* Pickers */}
                     <OptionPicker
@@ -239,4 +189,4 @@ const ReportBaggageModal = ({ isVisible, onClose, formData, reportBaggage, updat
     );
 };
 
-export default ReportBaggageModal;
+export default SearchBaggageModal;
