+++ mobile/package.json
@@ -38,8 +38,10 @@
     "react-native-reanimated": "~4.1.1",
     "react-native-safe-area-context": "~5.6.0",
     "react-native-screens": "~4.16.0",
+    "react-native-toast-message": "^2.3.3",
     "react-native-web": "~0.21.0",
     "react-native-worklets": "0.5.1",
+    "react-toastify": "^11.0.5",
     "tailwindcss": "^3.4.18"
   },
   "devDependencies": {
