+++ mobile/utils/string.utility.ts
@@ -1,0 +1,7 @@
+export const toTitleCase = (str: string): string => {
+    if (!str) return '';
+    return str.replace(
+        /\w\S*/g,
+        (txt) => txt.charAt(0).toUpperCase() + txt.substr(1).toLowerCase()
+    );
+};
